Added tweaked indent preferences
I noticed that the indentation of repeat…end blocks was not behaving
sensibly because it doesn't seem to work with various idioms - i.e.
"repeat x until y" was adding two indents when it shouldn't really add
any at all. I think it is better to leave the indent up to the
programmer in this case.